refactor(bolt): narrow bytecode plan module name
Rename the Stage 6 bytecode read-RAF planning module away from the overbroad verifier_plan name.

This keeps the new planning seam honest: it contains only Stage 6 bytecode read-RAF plan data, not the verifier planning layer as a whole.
