+++ crates/bolt/src/protocols/jolt/mod.rs
@@ -3,8 +3,8 @@ pub mod emit;
 pub mod oracles;
 pub mod params;
 pub mod phases;
+pub(crate) mod stage6_bytecode_read_raf_plan;
 pub mod validate;
-pub(crate) mod verifier_plan;
 
 pub use artifacts::{
     assemble_jolt_generated_crates, assemble_jolt_workspace_generated_crates, jolt_artifact_config,
