+++ crates/bolt/src/protocols/jolt/stage6_bytecode_read_raf_plan.rs
